refactor: split file search module
