+++ packages/app/src-tauri/src/file_search/tests.rs
@@ -1,0 +1,129 @@
+use std::fs;
+
+use rusqlite::Connection;
+
+use super::db::{indexed_path_exists, initialize_index_db, save_index_batch};
+use super::path_utils::is_path_allowed_by_options;
+use super::search::{parse_terms, search_index_db_with_connection, search_index_db_with_trigram};
+use super::types::{FileSearchOptions, IndexedFile, SEARCH_CANDIDATE_LIMIT};
+
+fn indexed_file(name: &str, path: &str, extension: &str) -> IndexedFile {
+    IndexedFile {
+        name: name.to_string(),
+        path: path.to_string(),
+        parent: "C:/root/docs".to_string(),
+        extension: extension.to_string(),
+        is_dir: false,
+    }
+}
+
+fn setup_search_db() -> Connection {
+    let mut connection = Connection::open_in_memory().unwrap();
+    initialize_index_db(&connection, true).unwrap();
+    save_index_batch(
+        &mut connection,
+        "C:/root",
+        "scan-test",
+        &[
+            indexed_file("ProjectNotes.md", "C:/root/docs/ProjectNotes.md", ".md"),
+            indexed_file("image.png", "C:/root/docs/image.png", ".png"),
+        ],
+    )
+    .unwrap();
+    connection
+}
+
+#[test]
+fn search_uses_fts_prefix_matches() {
+    let connection = setup_search_db();
+    let results = search_index_db_with_connection(
+        &connection,
+        "C:/root",
+        &parse_terms("proj"),
+        SEARCH_CANDIDATE_LIMIT,
+    )
+    .unwrap();
+
+    assert_eq!(results.len(), 1);
+    assert_eq!(results[0].name, "ProjectNotes.md");
+}
+
+#[test]
+fn search_falls_back_to_like_for_substring_matches() {
+    let connection = setup_search_db();
+    let results = search_index_db_with_connection(
+        &connection,
+        "C:/root",
+        &parse_terms("ject"),
+        SEARCH_CANDIDATE_LIMIT,
+    )
+    .unwrap();
+
+    assert_eq!(results.len(), 1);
+    assert_eq!(results[0].name, "ProjectNotes.md");
+}
+
+#[test]
+fn trigram_search_matches_filename_substrings() {
+    let connection = setup_search_db();
+    let results = search_index_db_with_trigram(
+        &connection,
+        "C:/root",
+        &parse_terms("ject"),
+        SEARCH_CANDIDATE_LIMIT,
+    )
+    .unwrap();
+
+    assert_eq!(results.len(), 1);
+    assert_eq!(results[0].name, "ProjectNotes.md");
+}
+
+#[test]
+fn search_combines_fts_and_extension_filters() {
+    let connection = setup_search_db();
+    let results = search_index_db_with_connection(
+        &connection,
+        "C:/root",
+        &parse_terms("proj .md"),
+        SEARCH_CANDIDATE_LIMIT,
+    )
+    .unwrap();
+
+    assert_eq!(results.len(), 1);
+    assert_eq!(results[0].name, "ProjectNotes.md");
+}
+
+#[test]
+fn indexed_path_exists_checks_exact_index_entries() {
+    let connection = setup_search_db();
+
+    assert!(indexed_path_exists(&connection, "C:/root/docs/ProjectNotes.md").unwrap());
+    assert!(!indexed_path_exists(&connection, "C:/root/docs/missing.md").unwrap());
+}
+
+#[test]
+fn path_allowed_by_options_accepts_roots_and_rejects_excludes() {
+    let temp_dir = tempfile::tempdir().unwrap();
+    let root = temp_dir.path().join("root");
+    let excluded = root.join("excluded");
+    let allowed_file = root.join("allowed.txt");
+    let excluded_file = excluded.join("blocked.txt");
+    fs::create_dir_all(&excluded).unwrap();
+    fs::write(&allowed_file, "allowed").unwrap();
+    fs::write(&excluded_file, "blocked").unwrap();
+
+    let options = FileSearchOptions {
+        roots: vec![root],
+        excluded_paths: vec![excluded],
+        include_hidden: false,
+    };
+
+    assert!(is_path_allowed_by_options(
+        &fs::canonicalize(allowed_file).unwrap(),
+        &options
+    ));
+    assert!(!is_path_allowed_by_options(
+        &fs::canonicalize(excluded_file).unwrap(),
+        &options
+    ));
+}
